Handle tree structures in JSONRenderer
This patch does two main things.  First, it allows the JSONRenderer to be used for
TreeGrids with a depth greater than one.  Second, it changes the JSON output to be
traditional JSON-style key-value pairs instead of separating column names from the
values.  As a result, the HTMLRenderer no longer relies on the JSONRenderer.