Add option for physical address KASLR shift
Starting with Linux kernel 4.8 KASLR introduced additional and
indepent[1] randomizations of not only the virtual kernel addresses but
also for the physical location where the various pieces of the kernel
are located.
Currently Volatility will try to determine the virtual address shift
by locating a certain string in physical memory, then calculate an
offset from that address ("swapper_address"), subtracting a kernel
identity mapping shift (in case of x64 this is 0xffffffff80000000) from
the "init_level4_gpt" virtual address as found in the profile's
System.map and then subtract this from "swapper_address":
shift_address = swapper_address - (init_level4_gpt - 0xffffffff80000000)
However the result, shift_address, only corresponds to the correct
virtual shift if 2 conditions hold true:
1. The physical shift is the same as the virtual shift.
2. The identity mapping shift is indeed 0xffffffff80000000
Starting with kernel 4.8 and KASLR enabled assumption 1 is no longer
true and the physical shift might is not be identical to the virtual
shift.
This commit introduces an option to pass a physical shift in addition
to the --shift option in order to validate Intel address spaces.
In a later commit a plugin is included to calculate virtual and
physical shift, which can be passed with --physicalshift and --shift.
[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=8391c73c96f28d4e8c40fd401fd0c9c04391b44a