Add plugin for KASLR shift detection
For Linux kernels >= 4.8 with KASLR enabled the standard DTB search
in plugins/overlays/linux/linux.py might not be successful or the
address space validation will fail. This plugin will try to find the
"correct" DTB candidate and derive the correct physical and virtual
ASLR shift which can then be passed to future Volatility invocations.
The DTB search is the same as in the Linux overlay, however it will
not just process the first possible DTB but will also look for other
candidates.
For each of those candidates all reachable PTEs will be collected and
for the PTEs pointing to the init_task physical address the
corresponding offsets in the PML4, PDPT, PD and PT that were taken in
order to reach that PTE  are compared to the ones expected from the
System.map virtual address for init_task.
The plugin's output looks like this:
DTB                Virtual Shift      Physical Shift
------------------ ------------------ ------------------
0x000000001a807000 0x000000003d800000 0x000000024c00000
This can be passed to vol.py:
./vol.py --shift=0x3d800000 --physicalshift=0x24c00000
For negative shift values the hex values need to be converted to
decimal in order to avoid confusion during option parsing.