Support for Yara Compiled Rules in Yarascan
Malfind assumes that the yara rule being used is a single text file and then compiles them.  This patch simply checks to see if the rule file is an already compiled rule and will load that rather than attempting to compile the rule.  If the rule is not a compiled rule, it will continue to compile as it did before.

This becomes very useful when compiling a directory of yara rules together into a single rule so volatility only needs to be run once, rather than creating a loop that will iterate over each rule and run/compile/scan, etc...